% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlp_survey}
\alias{nlp_survey}
\title{A Simulated Data of NLP Voters Survey}
\format{
A data frame with columns:

\describe{
\item{age}{Age group}
\item{gender}{Gender}
\item{vote_for}{Party voted for in the 2016 election}
\item{highest_educ}{Highest grade of education completed}
\item{state}{State}
\item{y}{Preference for NLP in 2020 election}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in NeverlandVoters.R
}
\usage{
nlp_survey
}
\description{
This is a simulated survey of voter preferences for the Neverland Labor Party
(NLP) that also includes several demographic variables. We use this data
throughout the documentation to demonstrate the usage of the package.
}
\examples{

 nlp_survey

}
\keyword{datasets}
